agent-ods/OmniParser/requirements-ods.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.0

transformers==4.45.2
safetensors==0.4.3
huggingface_hub==0.25.2
tokenizers==0.20.0

numpy==1.26.4
opencv-python==4.8.1.78
Pillow<12.0
ultralytics==8.3.0
requests==2.32.3
openai
easyocr
paddleocr
